refactored rest.scala to multiple files
